inheritance and implementation

+++ converter/src/main/java/com/handong/oh318/Point.java
@@ -2,22 +2,22 @@ package com.handong.oh318;
 
 public class Point {
     
-    private int x ; 
-    private int y ; 
+    private float x ; 
+    private float y ; 
 
-    public void setX(int x) { 
+    public void setX(float x) { 
         this.x = x; 
     }
 
-    public void setY(int y) { 
+    public void setY(float y) { 
         this.y = y ;
     }
 
-    public int getX() { 
+    public float getX() { 
         return x ;
     }
     
-    public int getY() { 
+    public float getY() { 
         return y ;
     }
 }
